Decided to include player into the game instead of having it as its own object. Added player and play functions to game and transferred the tests to game.spec.
