+++ spec/game.spec.js
@@ -1,11 +1,11 @@
 import Game from 'game';
-import Player from 'player';
+// import Player from 'player';
 import Board from 'board';
 
 describe('Game', function() {
   var testGame;
-      // testBoard,
-      // testPlayer;
+  // testBoard,
+  // testGame;
 
   beforeEach(function() {
     testGame = new Game();
@@ -13,18 +13,56 @@ describe('Game', function() {
     // testPlayer = new Player();
   });
 
-  describe('won', function() {
-    it('should check for a win if either player has won', function() {
-      testGame.player.players("Beylul", "Yeni");
-      // console.log(testPlayer.player1);
-      testGame.player.play(1, 0);
-      testGame.player.play(0, 1);
-      testGame.player.play(1, 1);
-      testGame.player.play(2, 1);
-      testGame.player.play(1, 2);
-      console.log(testGame.player.board.board);
-      expect(testGame.won()).toEqual("Beylul has won!");
+  describe('players', function() {
+    it('should assign both players', function() {
+      testGame.players("Beylul", "Yeni");
+      expect(testGame.player1).toEqual("Beylul");
+      expect(testGame.player2).toEqual("Yeni");
+    });
+  });
+
+  describe('play', function() {
+    it('should assign the right mark and player', function() {
+      testGame.players("Beylul", "Yeni");
+      testGame.play(0, 1);
+      expect(testGame.player).toBe("Beylul");
+      expect(testGame.mark).toEqual("x");
     });
 
+    it('should fill the spot with the designated mark, if empty', function() {
+      testGame.players("Beylul", "Yeni");
+      expect(testGame.play(0, 1)).toBe("Beylul picked spot [0][1]");
+    });
+
+    it('should increment counter by one', function() {
+      testGame.players("Beylul", "Yeni");
+      testGame.play(0, 1);
+      testGame.play(1, 1);
+      expect(testGame.counter).toEqual(3);
+    });
+
+    it('should ask the same player to go again if spot filled ', function() {
+      var check = testGame.check;
+      testGame.players("Beylul", "Yeni");
+      testGame.play(1, 1);
+      testGame.play(0, 0);
+      expect(testGame.play(0, 0)).toBe("Spot taken! Beylul try again.");
+      expect(testGame.counter).toEqual(3);
+    });
+
+    describe('won', function() {
+      it('should check for a win if either player has won', function() {
+        testGame.players("Beylul", "Yeni");
+        // console.log(testGame.player1);
+        testGame.play(1, 0);
+        testGame.play(0, 1);
+        testGame.play(1, 1);
+        testGame.play(2, 1);
+        testGame.play(1, 2);
+        console.log(testGame.board);
+        expect(testGame.won()).toEqual("Beylul has won!");
+      });
+
+    });
   });
 });
