+++ src/game.js
@@ -1,10 +1,43 @@
-
-import Player from 'player';
+import Board from 'board';
 
 var Game = function() {
-  this.player = new Player();
-  // this.board = new Board();
+  this.board = new Board();
 };
+
+Game.prototype.players = function (player1, player2) {
+  this.player1 = player1;
+  this.player2 = player2;
+};
+
+Game.prototype.play = function (row,column) {
+  //if counter == 9 invoke won function and clear board
+  if (this.counter === undefined){
+    this.counter = 1;
+  }
+
+  if(this.counter % 2 == 1) {
+    this.player = this.player1;
+    this.mark = "x";
+  }
+  else {
+    this.player = this.player2;
+    this.mark = "o";
+  }
+
+  var isFilled  =  this.board.filled(row,column);
+  var reply;
+  if(isFilled === false){
+    this.board.fill(row, column, this.mark);
+    reply = this.player + " picked spot " + "[" + row + "]" + "[" + column + "]";
+    this.counter += 1;
+  }
+  else {
+    reply = "Spot taken! " + this.player + " try again.";
+  }
+  // this.board.hasWon(this.counter);
+  return reply;
+};
+
 
 //move won into board
 Game.prototype.won = function () {
@@ -16,10 +49,10 @@ Game.prototype.won = function () {
   this.reply = "";
 
   //horizontal
-  for (var index = 0; index < this.player.board.board.length; index++) {
-    if (this.player.board.board[index][0] === this.player.board.board[index][1] && this.player.board.board[index][1] === this.player.board.board[index][2] && this.player.board.board[index][0] !== undefined) {
+  for (var index = 0; index < this.board.board.length; index++) {
+    if (this.board.board[index][0] === this.board.board[index][1] && this.board.board[index][1] === this.board.board[index][2] && this.board.board[index][0] !== undefined) {
         this.won = true;
-        this.mark = this.player.board.board[index][0];
+        this.mark = this.board.board[index][0];
       }
   }
 
@@ -29,10 +62,10 @@ Game.prototype.won = function () {
 
   if (this.won === true) {
     if (this.mark === 'x') {
-      this.reply = this.player.player1 + " has won!";
+      this.reply = this.player1 + " has won!";
     }
     else if (this.mark === 'o') {
-      this.reply = this.player.player2 + " has won!";
+      this.reply = this.player2 + " has won!";
     }
     return this.reply;
   }
